Simplified code which handles calculation of anchor points.
The code to calculate coordinates and anchor points has been consolidated in a single function "E:CalculateMoverPoints(mover, nudgeX, nudgeY)".

Nudging anchors was suffering from a bug which caused the anchor to skip to the edge of the screen when it crossed the center. The new simplified version fixes this issue as well.