Added .unitframeType element to all created unitframes, so we can easily identify the type without having to string.find the name.
This allows external plugins to only execute code for certain unitframe types when they hook the Configure_ functions for unitframe elements. Using frame.unit is not always precise, as a raid unitframe may have "player" as unit, and searching the name of the unitframe is inefficient.

Example of how you can choose to only update specific elements on specific frames:
	hooksecurefunc(UF, "Configure_Power", function(self, frame)
		if frame.unitframeType == "player" or frame.unitframeType == "target" then
			--Do something that only affects power bar on player or target frame
		end
	end)