Fixed issue which caused unitframe health deficit to draw above all other elements
This was caused by portrait.overlay frame having a frame level of 10000 due to it being set to negative value (player unitframe frame level is 2, and then 5 was subtracted).

For some reason any attempt to set a higher frame level on the unitframe itself will have no effect.

This fixes Elv/elvui-beta#9