Fixed E:GetScreenQuadrant so it divides the screen into 9 quadrants instead of 16.
Previously 1 quadrant was assigned to left and right, while center was assigned 2 quadrants. The same was true for top and bottom.

This didn't really give a realistic picture of where a frame was positioned, as it had to be positioned very near the edge of the screen to not be considered as "center".