revert TalentMicroButtonAlert skin functionality to former state. this closes #299 (reverted commit ee270117)
reverting this will force skin/place the alert if we have the option enabled to show it (with no means to remove the skin, having the option disabled kills the frame). its like this again to prevent tickets since the microbar can be moved to where this frame is off-screen, it should prevent further tickets created on this.